Move timeline things from catalog.rs to a new file
